// src/lib/types.ts
export interface ReturnData {
  date: string; // Changed from CumulativeReturn
  value: number;
}